requirements.txt
openai==0.27.8
python-dotenv==1.0.0
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
markdown==3.4.3
weasyprint==52.5
pydyf==0.5.0
